requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via httpx
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via pytfe
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via pytfe
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
pydantic==2.12.5
    # via pytfe
pydantic-core==2.41.5
    # via pydantic
pynacl==1.5.0
    # via -r requirements.in
pytfe==0.1.2
    # via -r requirements.in
python-dotenv==1.2.2
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   pytfe
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
